Add Integration testing for APIView
