+++ src/dotnet/APIView/APIView.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29021.251
+# Visual Studio Version 17
+VisualStudioVersion = 17.2.32602.215
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "APIView", "APIView\APIView.csproj", "{F3B9015F-4C33-4861-A3D6-E6AAE4C7772A}"
 EndProject
@@ -17,6 +17,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 	ProjectSection(SolutionItems) = preProject
 		.editorconfig = .editorconfig
 	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "APIViewIntegrationTests", "APIViewIntegrationTests\APIViewIntegrationTests.csproj", "{8C53BC0C-DEEC-4810-A0D8-A232A25BA4B7}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -44,6 +46,10 @@ Global
 		{09661CDE-E40B-4892-86DC-6DD79A279220}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{09661CDE-E40B-4892-86DC-6DD79A279220}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{09661CDE-E40B-4892-86DC-6DD79A279220}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8C53BC0C-DEEC-4810-A0D8-A232A25BA4B7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8C53BC0C-DEEC-4810-A0D8-A232A25BA4B7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8C53BC0C-DEEC-4810-A0D8-A232A25BA4B7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8C53BC0C-DEEC-4810-A0D8-A232A25BA4B7}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
