+++ src/dotnet/APIView/APIViewTest/APIViewTest.csproj
@@ -6,7 +6,6 @@
     <IsPackable>false</IsPackable>
     <PreserveCompilationContext>true</PreserveCompilationContext>
     <EmbeddedResourceUseDependentUponConvention>false</EmbeddedResourceUseDependentUponConvention>
-    <UserSecretsId>581d94f0-bdc6-4bde-92cb-06e91efbcdba</UserSecretsId>
   </PropertyGroup>
 
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
@@ -20,7 +19,6 @@
 
   <ItemGroup>
     <PackageReference Include="Microsoft.Azure.Cosmos" Version="3.26.1" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.UserSecrets" Version="3.1.13" />
     <PackageReference Include="Microsoft.Extensions.DependencyModel" Version="2.1.0" />
     <PackageReference Include="Azure.Storage.Blobs" Version="1.0.0-preview.1" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.2.0" />
